Fix issue #8152662: Unable to share a contact via Email
When contacts provider internally recurses into itself, we need
to clear the binder calling identity before doing so, so that it
is consistent with the calling package name we will now be using.

Note that this also means that we won't be doing any further
permission checks against the original caller for these internal
provider calls...  which is probably more correct anyway, because
we've already done the permission check on the original incoming
URI and we wouldn't want that to now fail due to some permission
strangeness on another URI it is using internally.

Change-Id: Icc3903287f91f35959a80168a6c107b1803f93a8