Fixed issue where some contacts are not added to delete log.
Contacts with multiple raw contacts were not being inserted into the delete
log when accounts are removed.  This was due to the raw contacts being removed
in batch instead of one at a time. The logic to determine whether a contact
can be deleted was determining whether other raw contacts exist.  This does
not work when all raw contacts are removed in a single batch.

Added logic to pre-select contacts that will be deleted.

Bug: 8696462
Change-Id: I95adccf9e6756bbf6ca9dd7d144c1d9ee8905631