Do not ignore leading "*"s in phone lookups
After a phone lookup, "remove" entries from the cursor that
don't match the queried number w.r.t. to whether it has a leading
"*". "Removing" entries from a cursor involves making a partial
copy of the cursor.

This CL affects as little functionality as possible, so that it
can make it into MR2. Ie, we continue to ignore leading "*"s
in other CP2 queries.

MANUAL TESTS:
-Add a "*123" Emergency contact and "123" Voicemail contact
-Call "*123", then call "123". You should see two seperate
 entries in your call log. In addition, the incall screens
 should show Emergency and Voicemail respectively (although
 the "123" call doesn't work properly inside the incall
 UI right now. This is a seperate bug b/13195342).

Bug: 13195334
Change-Id: I7e1938ccb085630072611eb63148c5714b62fee8