Update CP2 snippeting argument defaults (1/2)
The Contacts app was the only user of the snippeting functionality in
CP2, so update the default snippeting arguments to those in the Contacts
app.

Also fix some tests that had incorrect assumptions about outdated
functionality.

There is one remaining broken test in CP2 - testSearchByEmail. That has to do
with some special treatment with indexing email addresses and FTS search which
will take some more time to figure out.

Bug: 6948882
Change-Id: Id9af3ceb2dfa43a11ae528e3ab6b42371130a1f7