Upgrade calllog DB to use ICC ID & source_phone_number.
To accomodate backup and restore of the call log, we have to convert the
existing sub-ID to an ICC ID.  Also, add the source_phone_number to the
list of call log columns.

The ICC ID improves backup and restore because it is persistent across
different devices where a sub ID is not.  If the user completely changes
their SIM card, the source_phone_number will be helpful in identifying
that scenario if the user kept their old phone.

Change-Id: Ib35e25216691e580b5ea539929ff16b2e3d30b88