Enterprise quick contact 2/2
Now openQuickContact goes thorough DPM.  When a lookup URI is build with
a lookup key returned by the enterprise lookup APIs for a corp contact, the
lookup key will have a special prefix.  DPM checks it and launches QC on
the work profile, if the policy allows it.

Design doc: go/cp2-mnc-enterprise-dd

Bug 19546108

Change-Id: Iccba8254269180281af0154f4b9c20cbfff5c27d