Re-calculate hash_id for photo field.
In the old logic, we use data15 to compute the hash_id for a photo.
But data15 values are not the same for the original value and synced
value from the server, because there're some different compressing method
used in CP2 and focus sync adapter.
Because there's only one photo for each raw_contact, in the current cl
we only used photo mimetype to generate hash_id.

BUG 25955017

Change-Id: I0cdc8583f6f32ea2dab69de241ad2c6c05332319