[FBE] Introduce shadow calllog provider [2/2]
Introduce a new provider that's a clone of the calllog provider but
is EA.

The provider is hidden and requires MANAGE_USERS to access, and responsible
for storing calllog entries that are inserted when the real provider is still
encrypted.  When the real provider starts, it copies the entries from the shadow
as well as user-0's real provider, and clears the shadow.

Also fix b/26516259

Bug 26183949

Change-Id: If44b46709e2e7b1651b41c09d900e1cb2777dc56