Use LocaleList to support multi-locale to improve name "bucketing".
We used to keep the previous locale on the locale change event
and used it as the secondary locale, but now that the framework
officially supports "LocaleList", we switch to the new mechanism.

On the contacts app, this affects the name "buckets".  Adding more
locales means you may get more buckets for the letters specific to
these locales.

- Now ContactLocaleUtilsTest doesn't check supported collators.
Turned out quite a few collators were considered missing (because now
they're tied to languages, not language-country pares?) and were
simply skipped.  In the meantime, apparently the labels for some
languages have changed.

Bug 26142164

Change-Id: I4ff7e7d51d000adb6b996eb77da07af5cd44849f