Removing mimetype and package cache
We need to remove caches since after enabling WAL, we will have writers
work in non-exclusive transactions (b/31600627) which might lead to
readers reading uncommitted values if the cache is updated and the
transaction fails.
Test: Added a test testGetCommonMimeTypeIds to test the prepopulated
cache for common mimetypes added in this change. Verified that existing
tests in ContactsProvider2Test and ContactsDatabaseHelperTest pass.

Bug: 31599660
Change-Id: I804ae9e952029a7f1b702347f866cb5e5d951f3e