Add voicemail backup/restore columns
BACKED_UP and RESTORED are required for dialer voicemail backup/restore
feature for N MR2. As there are no API bump the fields will be hidden
in NMR2. These fields will be made public in O.

ARCHIVED and IS_OMTP_VOICEMAIL is not used for NMR2, but added in
advance to avoid multiple database upgrades.

Test: VoicemailProviderTest
Bug: 34463609
Change-Id: I95ef21950337ca1f8bc79cc0493f7c302ec05617