Allow voicemail DIRTY flag to be explicity set
The DIRTY flag is used to indicate there are local changes that needs
to be uploaded to the server. Previously this flag is automatically set
to 1 if the modifying package does not own the voicemail. For example
when dialer marked the voicemail as read, ContactsProvider will mark
it as dirty, and the source telephony will query for dirty entries
and upload it to the server.

After visual voicemail is unbundled, the voicemail source is no longer
guaranteed to be separated from the voicemail UI.

In this CL, the DIRTY field will no longer be overriden if the caller
sets the field.

Bug: 38302993
Test: CtsProviderTestCases

Change-Id: I4a3d08a7ab2b9d61d271ae548ecdbc358af48f3e