Implement DIRTY_RETAIN constant
Previously the DIRTY flag can only be automatically or explicitly set.
There are no option to express "This operation does not change the DIRTY
flag". This causes issue when the voicemail source is downloading
changes from the server instead of uploading and would clobber the flag.

After this CL, if DIRTY is updated with DIRTY_RETAIN, the flag will not
be updated for the operation.

Change-Id: I23a23a11f7d2177d10b65bbc33875b7eb838f7c1
Fixes: 64371667
Test: CtsProviderTestCases#VoicemailContractTest