Fix issue with failing ContactsProviderTest.
We used to call DefaultDialerManager.isDefaultOrSystemDialer; that method
has an explicit empty/null check for the callingPackage, which I didn't
replicate here.

Honestly the test probably should have a null callingPackage, but in the
interest of maintaining the same code/test behavior, changing this method
works to solve the problem.

Test: Re-run unit tests
Bug: 144004289
Change-Id: I72e3b379ca13db3860f9882585f9f3d06daa1bb5