Grant READ_PRIVILEGED_PHONE_STATE to contacts provider to access ICC ID
In Android 10 access to device identifiers was moved from a runtime
permission to the privileged READ_PRIVILEGED_PHONE_STATE permission.
One of the non-resettable identifiers missed was SubscriptionInfo#
getIccId. The contacts provider currently uses this ID to upgrade
to version 1003 of the database; to ensure there are no issues with
this upgrade this permission is being granted to the contacts
provider.

Bug: 131909991
Test: atest ContactsProviderTests
Change-Id: Iafd9d036993a2d8d765d9ca001c91076d8df18b0