Ensure update/deletes in the call log notify content observers.
Update/delete in CallLogProvider used to use DbModifierWithNotification
which took care of this.  Due to a change made to CallLogProvider to use
a query builder class, DbModifierWithNotification was not being used any
more.

Made the notifyCallLogChange accessible as a static method and also use it
in CallLogProvider on update/delete.

There is a bunch of other logic in DbModifierWithNotification in the
update method that needs to be migrated over, however this change fixes
the immediate bug.

Test: Manual test; clear the call log and verify Dialer recents lists
updates itself immediately.
Test: Manual test: clear a single row in the call log and verify dialer
recent list updates itself immediately.
Test: Added update/delete CTS tests that verify content observer triggers
for update and delete.
Fixes: 174243006
Bug: 174211399

Change-Id: Ie716d3d739c6854d7c40a1f94498bff15931b272