Fix bug where CP2 tries to access a closed cursor
Patchset 5 of ag/20470472 used try-with-resources in queryMergedContacts.
That introduced the issue in b/263386251#comment2. The problem is that
the cursor is closed even when there is no exception, which means that
the method never returns a useful cursor.
This CL returns to the version of the code from previous patchsets.

Bug: 263386251
Test: adb shell content query --uri content://com.android.contacts/contacts/enterprise
Change-Id: I5bf688446e399eb9a15ffea20d06af4dad5fd594