Restrict contact access to launchable cloned apps
Earlier, we used the cloneable_apps.xml as an allowlist to restrict
cloned apps to access cross-profile contacts. This made it difficult to
maintain CTS tests for clone contacts provider as the allowlist is empty
by default. This change moves the allowlisting logic to use package
manager call to identify launchable activities and use them to
restrict access to cross-profile contacts.

The mLaunchableCloneAppsCache is also added to ensure the results for
a package are cached. Since, app uids are used as a key every cache
entry, the entries are refreshed after 60s to ensure the uid values are
not stale. The cache is also cleaned up once every 7 days to remove
outdated entries using additional space (apps that have been removed).

Bug: 271752108
Test: atest CloneContactsProvider2Test - added a couple of test case
corresponding to the app allowlist check
Change-Id: I6af0a69ac0c84a85e6cb39b2d573dec78a050a81

Change-Id: Id7282c9bf9b58fd32cf64cb0842d5e34c9ca2f18