Correct broadcast action for ContactsProvider2 receiver
The runtime receiver in ContactsProvider2 was reported as part of
a lint check due to registering for a non-system broadcast. The
action for which it is registering is expected as an extra in
the ACTION_PHONE_ACCOUNT_REGISTERED broadcast; since the receiver
in this class has a check for this action, this commit updates the
IntentFilter for this receiver to use this broadcast action instead.

Bug: 234659204
Test: ContactsProvider2Test
Change-Id: I6d3591521243f2932f5b6ef8a0f35ab894bdd331