Refactored updateCustomContactVisibility in ContactsDatabaseHelper
The method updateCustomContactVisibility was spending 2.47% of total
userspace CPU time parsing SQL queries during the downloadContacts CUJ.
The main reason of this is the use of a dynamic string which has the
optionalContactId in it, instead of being passed as a selection
argument.

Android's SQLite layer has the PreparedStatementCache, which is a LRU
cache mapping a query string to an SQLite prepared statement. When a SQL
query produces a cache hit, we avoid SQLite's query compilation step. By
passing a different optionalContactId inside the string on every contact
insert the cache is missed and compilation/parsing is requried.

To avoid this, the optionalContactId is replaced with ? in the SQL query
and the contact id is passed as a selection arg. Also, removing the
string concatenation with static variables makes the SQL query much more
readable and is much less prone to error.

Test: atest ContactsProviderTests CtsContactsProviderTestCases
Test: atest ContactsDatabaseHelperTest
Flag: EXEMPT refactor
Bug: 348457141
Bug: 350056429
Change-Id: I06181019092b692cbb43e4918e57fde341cc840b