Update search_index table on account removal and contact aggregation
When an account is removed all of its data is removed including its
contacts. Previously in removeDataOfAccount the contacts were removed
from the raw_contacts table, but not from the search_index table.
Refactored the method such that the deleted contacts are removed from
the search index and contacts who had a raw contact removed are updated
in the search index.

Too handle the case where there is an excessive amount of contacts (>5K)
which need to be deleted from the search index, we simply clear the
search_index table and rebuild it.

Also, updated the contacts aggregation logic such that it removes
obsolete contact ids from the search index once a raw contact is
updated with the contact id of the new aggregated contact. Previously,
the aggregator would only delete the contact id from the contacts table
but wouldn't update the search index table. This resulted in duplicating
many rows in the search index.

Also, updated ContactsProvider2Test to use AndroidJUnit4 test runner in
order to use feature flag annotations.

Test: atest ContactsProviderTests
Bug: 363260703
Flag: com.android.providers.contacts.flags.cp2_sync_search_index_flag
Change-Id: I9992ad02bae68bc46ad3b2f5aababbde091ca5d1